refactor(channels): simplify rail context
