+++ apps/web/src/features/channels-view/components/ChannelsMobileView.tsx
@@ -58,10 +58,10 @@ export function ChannelsMobileView(props: {
     DEBUG_SETTING_KEYS.FORCE_EMPTY_STATES
   );
   const listId = createUniqueId();
-  const { callActivity, incomingCallIds, callStatuses } = useChannelCalls();
+  const channelCalls = useChannelCalls();
   const channelActivity = useChannelRailActivity(
     () => props.channels,
-    callActivity
+    channelCalls
   );
 
   const visibleChannels = createMemo(() =>
@@ -183,8 +183,10 @@ export function ChannelsMobileView(props: {
                       currentUserId()
                     )}
                     unread={channelActivity.unreadChannelIds().has(channel.id)}
-                    callStatus={callStatuses().get(channel.id)}
-                    incomingCallId={incomingCallIds().get(channel.id)}
+                    callStatus={channelActivity.callStatuses().get(channel.id)}
+                    incomingCallId={channelActivity
+                      .incomingCallIds()
+                      .get(channel.id)}
                     selected={state.selectedChannelId === channel.id}
                     focused={false}
                     onActivate={() => openChannel(channel)}
